+++ sdk/formrecognizer/azure-ai-formrecognizer/src/test/java/com/azure/ai/formrecognizer/FormTrainingClientTest.java
@@ -170,8 +170,8 @@ public class FormTrainingClientTest extends FormTrainingClientTestBase {
     public void listCustomModels(HttpClient httpClient, FormRecognizerServiceVersion serviceVersion) {
         client = getFormTrainingClient(httpClient, serviceVersion);
         for (CustomFormModelInfo modelInfo : client.listCustomModels()) {
-            assertTrue(modelInfo.getModelId() != null && modelInfo.getRequestedOn() != null
-                && modelInfo.getCompletedOn() != null && modelInfo.getStatus() != null);
+            assertTrue(modelInfo.getModelId() != null && modelInfo.getTrainingStartedOn() != null
+                && modelInfo.getTrainingCompletedOn() != null && modelInfo.getStatus() != null);
         }
     }
 
@@ -183,8 +183,8 @@ public class FormTrainingClientTest extends FormTrainingClientTestBase {
     public void listCustomModelsWithContext(HttpClient httpClient, FormRecognizerServiceVersion serviceVersion) {
         client = getFormTrainingClient(httpClient, serviceVersion);
         for (CustomFormModelInfo modelInfo : client.listCustomModels(Context.NONE)) {
-            assertTrue(modelInfo.getModelId() != null && modelInfo.getRequestedOn() != null
-                && modelInfo.getCompletedOn() != null && modelInfo.getStatus() != null);
+            assertTrue(modelInfo.getModelId() != null && modelInfo.getTrainingStartedOn() != null
+                && modelInfo.getTrainingCompletedOn() != null && modelInfo.getStatus() != null);
         }
     }
 
@@ -250,8 +250,8 @@ public class FormTrainingClientTest extends FormTrainingClientTestBase {
                     CustomFormModelInfo> copyPoller = client.beginCopyModel(actualModel.getModelId(), target);
                 CustomFormModelInfo copyModel = copyPoller.getFinalResult();
                 assertEquals(target.getModelId(), copyModel.getModelId());
-                assertNotNull(actualModel.getRequestedOn());
-                assertNotNull(actualModel.getCompletedOn());
+                assertNotNull(actualModel.getTrainingStartedOn());
+                assertNotNull(actualModel.getTrainingCompletedOn());
                 assertEquals(CustomFormModelStatus.READY, copyModel.getStatus());
             });
         });